requirements.txt
flask==3.0.0
flask-cors==4.0.0
bcrypt==4.1.2
pyjwt==2.8.0
python-dotenv==1.0.0
firebase-admin==6.4.0
google-cloud-firestore==2.13.1
google-generativeai==0.3.2
langchain==0.3.0
langchain-google-genai==2.0.0